+++ RQ_RLMI_CPP/src/RLMINode.h
@@ -21,7 +21,7 @@ public:
     RLMINode(KVEntry **trainData, int _dataSize);
     std::vector<double> build();
     void calMuSig();
-    inline double predict(const unsigned int &key) const {
+    inline double predict(const unsigned long long int &key) const {
         double _key = (1.0 * key - this->mu) / this->sig;
         return this->_a * _key + this->_b;
     }
