+++ RQ_RLMI_CPP/src/Range.h
@@ -10,9 +10,9 @@ public:
     unsigned int id;
 
 
-    Range(unsigned int _Low, unsigned int _High, unsigned int _id);
+    Range(unsigned long long int _Low, unsigned long long int _High, unsigned long long int _id);
     void printSelf();
-    inline bool match(const unsigned int &pos) const {
+    inline bool match(const unsigned long long int &pos) const {
         if(pos >= this->Low && pos <= this->High) {
             return true;
         } else {
